@@ -4,17 +4,19 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "hybrid_crypto_project"
-version = "0.1.0"
-description = "Enterprise-grade Hybrid Cryptography CLI reference implementation"
-requires-python = ">=3.11"
+version = "1.0.0"
+description = "Production-grade hybrid cryptography system with mandatory post-quantum security"
+requires-python = ">=3.12"
 authors = [ { name = "hybrid team" } ]
 readme = "README.md"
 dependencies = [
-    "cryptography>=40.0",
+    "cryptography>=43.0.0",
+    "boto3>=1.36.0",
+    "typer>=0.15.1",
 ]
 
-[project.optional-dependencies]
-pq = ["python-oqs>=0.7"]
+[project.scripts]
+hybrid-crypto = "hybrid_crypto_project.main:main"
 
 [tool.setuptools.packages.find]
 where = ["."]
@@ -22,3 +24,14 @@ include = ["hybrid_crypto_project*"]
 
 [project.urls]
 Homepage = "https://example.invalid/"
+
+[tool.pytest.ini_options]
+pythonpath = ["."]
+testpaths = ["tests"]
+
+[tool.ruff]
+line-length = 100
+target-version = "py312"
+
+[tool.ruff.lint]
+select = ["E", "F", "I", "B", "UP", "N", "ARG"]
