@@ -5,8 +5,21 @@ module will call `cli.main()` from the packaged CLI implementation.
 """
 from __future__ import annotations
 
-from .cli import main
+
+def _run() -> int:
+    try:
+        from .cli import main
+    except ModuleNotFoundError as exc:
+        if exc.name == "typer":
+            print(
+                "Missing dependency: typer. "
+                "Run `make install` or `.venv/Scripts/pip install typer`."
+            )
+            return 1
+        raise
+
+    return main()
 
 
 if __name__ == "__main__":
-    main()
+    raise SystemExit(_run())
